+++ kernel/data-pipeline/dialect/mysql/src/main/java/org/apache/shardingsphere/data/pipeline/mysql/ingest/incremental/binlog/data/unsigned/impl/MySQLBinlogUnsignedBigintHandler.java
@@ -30,6 +30,6 @@ public final class MySQLBinlogUnsignedBigintHandler implements MySQLBinlogUnsign
     
     @Override
     public Number handle(final Long value) {
-        return value < 0L ? BIGINT_MODULO.add(BigInteger.valueOf(value)) : value;
+        return value < 0L ? BIGINT_MODULO.add(BigInteger.valueOf(value)) : BigInteger.valueOf(value);
     }
 }