+++ kernel/data-pipeline/dialect/mysql/src/main/java/org/apache/shardingsphere/data/pipeline/mysql/ingest/incremental/binlog/data/unsigned/impl/MySQLBinlogUnsignedIntHandler.java
@@ -28,6 +28,6 @@ public final class MySQLBinlogUnsignedIntHandler implements MySQLBinlogUnsignedN
     
     @Override
     public Number handle(final Integer value) {
-        return value < 0 ? INT_MODULO + value : value;
+        return value < 0 ? INT_MODULO + value : value.longValue();
     }
 }