+++ kernel/data-pipeline/dialect/mysql/src/test/java/org/apache/shardingsphere/data/pipeline/mysql/ingest/incremental/binlog/data/unsigned/impl/MySQLBinlogUnsignedBigintHandlerTest.java
@@ -22,6 +22,7 @@ import org.junit.jupiter.api.Test;
 import java.io.Serializable;
 import java.math.BigInteger;
 
+import static org.hamcrest.CoreMatchers.instanceOf;
 import static org.hamcrest.CoreMatchers.is;
 import static org.hamcrest.MatcherAssert.assertThat;
 
@@ -32,8 +33,10 @@ class MySQLBinlogUnsignedBigintHandlerTest {
     @Test
     void assertHandle() {
         Serializable actual = handler.handle(1L);
-        assertThat(actual, is(1L));
+        assertThat(actual, instanceOf(BigInteger.class));
+        assertThat(actual, is(new BigInteger("1")));
         actual = handler.handle(-1L);
+        assertThat(actual, instanceOf(BigInteger.class));
         assertThat(actual, is(new BigInteger("18446744073709551615")));
     }
 }