+++ kernel/data-pipeline/dialect/mysql/src/test/java/org/apache/shardingsphere/data/pipeline/mysql/ingest/incremental/binlog/data/unsigned/impl/MySQLBinlogUnsignedIntHandlerTest.java
@@ -21,6 +21,7 @@ import org.junit.jupiter.api.Test;
 
 import java.io.Serializable;
 
+import static org.hamcrest.CoreMatchers.instanceOf;
 import static org.hamcrest.CoreMatchers.is;
 import static org.hamcrest.MatcherAssert.assertThat;
 
@@ -31,8 +32,10 @@ class MySQLBinlogUnsignedIntHandlerTest {
     @Test
     void assertHandle() {
         Serializable actual = handler.handle(1);
+        assertThat(actual, instanceOf(Long.class));
         assertThat(actual, is(1L));
         actual = handler.handle(-1);
+        assertThat(actual, instanceOf(Long.class));
         assertThat(actual, is(4294967295L));
     }
 }