+++ kernel/data-pipeline/dialect/mysql/src/test/java/org/apache/shardingsphere/data/pipeline/mysql/ingest/incremental/binlog/data/unsigned/impl/MySQLBinlogUnsignedSmallintHandlerTest.java
@@ -21,6 +21,7 @@ import org.junit.jupiter.api.Test;
 
 import java.io.Serializable;
 
+import static org.hamcrest.CoreMatchers.instanceOf;
 import static org.hamcrest.CoreMatchers.is;
 import static org.hamcrest.MatcherAssert.assertThat;
 
@@ -31,8 +32,10 @@ class MySQLBinlogUnsignedSmallintHandlerTest {
     @Test
     void assertHandle() {
         Serializable actual = handler.handle((short) 1);
+        assertThat(actual, instanceOf(Integer.class));
         assertThat(actual, is(1));
         actual = handler.handle((short) -1);
+        assertThat(actual, instanceOf(Integer.class));
         assertThat(actual, is(65535));
     }
 }